I have edited the iteration code so that we are only looking into a small section of the curve resulting in an approximate linear log plot.
The gradient is then calculated using np.polyfit as a least squares fit
function was not found with a quick look through google. polyfit
worked fine with some basic examples of y = mx + c that we tested it with.

We have a problem that when we change the step size from 1 to anything
smaller the graph seems to behave erratically. I would appreciate it if you
could look into this and see if you can spot the problem.

For now with the constants in the code we get a gradient of approx
-1.6