me and scott have worked on dirac.py it now does... a thing.
We have added the grid spacing you used in iteration.py
We then used the formula for staggered fermions and stuck
that in for the differential part of the dirac eq.

The first d/dx isn't right as we did r[i +1]  r[i] but this
may not matter too much.

Later we will look into selecting the middle range of the
eigenvalues.