Faster version of RPA_intra.
Instead of integrating the RPA intraband response for each pair of
values (r, r'), one can find the response as a function of
distance between the two points in space, R = sqrt(r1**2 + ...).
Then, the response can be obtained by integrating this quantity
as a function of relative angle theta. For this integration,
it is sufficient to employ spline interpolation of the response
as a function of R. This spline interpolation has to be calculated
only once.